Create fragments for Canvas to reuse in multiple queries
