+++ RealifeTech-CoreSDK/GraphQL/GraphQLFiles/canvas_fragment_screen.graphql
@@ -1,0 +1,4 @@
+fragment screenTranslation on ScreenTranslation {
+  language
+  title
+}
