+++ RealifeTech-CoreSDK/GraphQL/GraphQLFiles/canvas_getWidgetsByScreenId.graphql
@@ -1,30 +1,9 @@
+#import "./canvas_fragment_widget.graphql"
+
 query getWidgetsByScreenId($screenId: ID!, $pageSize: Int!, $page: Int = 1) {
   getWidgetsByScreenId(screenId: $screenId, pageSize: $pageSize, page: $page) {
     edges {
-      id
-      style {
-        type
-        size
-        padded
-      }
-      viewAllUrl
-      widgetType
-      variation {
-        fetchType
-        contentIds
-        params {
-          key
-          value
-        }
-        engagementParams {
-          key
-          value
-        }
-        translations {
-          language
-          title
-        }
-      }
+      ... widget
     }
     nextPage
   }
